refacto SuiviManager and SuiviFactory #3052
