+++ src/EventSubscriber/InterventionCanceledSubscriber.php
@@ -43,14 +43,11 @@ class InterventionCanceledSubscriber implements EventSubscriberInterface
             $suivi = $this->suiviManager->createSuivi(
                 user: $currentUser,
                 signalement: $intervention->getSignalement(),
-                params: [
-                    'description' => $description,
-                    'type' => Suivi::TYPE_AUTO,
-                ],
+                description: $description,
+                type: Suivi::TYPE_AUTO,
                 isPublic: true,
                 context: Suivi::CONTEXT_INTERVENTION,
             );
-            $this->suiviManager->save($suivi);
 
             $this->visiteNotifier->notifyUsagers(
                 $intervention,
