+++ src/EventSubscriber/InterventionConfirmedSubscriber.php
@@ -66,14 +66,11 @@ class InterventionConfirmedSubscriber implements EventSubscriberInterface
             $suivi = $this->suiviManager->createSuivi(
                 user: $currentUser,
                 signalement: $intervention->getSignalement(),
-                params: [
-                    'description' => $description,
-                    'type' => Suivi::TYPE_AUTO,
-                ],
+                description: $description,
+                type: Suivi::TYPE_AUTO,
                 isPublic: $isUsagerNotified,
                 context: Suivi::CONTEXT_INTERVENTION,
             );
-            $this->suiviManager->save($suivi);
 
             if ($isUsagerNotified) {
                 $this->visiteNotifier->notifyUsagers($intervention, NotificationMailerType::TYPE_VISITE_CONFIRMED_TO_USAGER);
