+++ src/Messenger/MessageHandler/NewSignalementCheckFileMessageHandler.php
@@ -203,13 +203,10 @@ class NewSignalementCheckFileMessageHandler
         return $this->suiviManager->createSuivi(
             user: $userAdmin,
             signalement: $signalement,
-            params: [
-                'type' => Suivi::TYPE_AUTO,
-                'description' => $this->description,
-            ],
+            description: $this->description,
+            type: Suivi::TYPE_AUTO,
             isPublic: true,
             context: Suivi::CONTEXT_NOTIFY_USAGER_ONLY,
-            flush: true,
         );
     }
 }
